Now catching only ComboBox.ON_HIDDEN events.
Also attempting to cancel the edit when the user hits ESC. One thing
I had to do in order to facilitate this was to wait to commit by putting
the contents of ComboBox.ON_HIDDEN event handler in a Platform.runLater
so that it will run after the KeyEvents have processed. Otherwise,
the cell commits before we realize that the CheckComboBox closed
because of a ESC key.